+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added option to read classes for HTML pages from file
 ### Changed
+- changed the implementation of GA4GHClaimSource because the EGA Permissions service
+  changed its API to be GA4GH-compatible
 - Changed logging configuration - log all classes at the level specified while building
 - Consent uses the loading of HTML classes from the file for some elements
 - Updated templates for consent page